+++ libraries/api-client/src/conversation/conversationApi/index.ts
@@ -1,6 +1,6 @@
 /*
  * Wire
- * Copyright (C) 2021 Wire Swiss GmbH
+ * Copyright (C) 2022 Wire Swiss GmbH
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as published by
@@ -17,13 +17,4 @@
  *
  */
 
-import {QualifiedUserClients} from './QualifiedUserClients';
-
-export interface MessageSendingStatus {
-  deleted: QualifiedUserClients;
-  failed_to_confirm_clients?: QualifiedUserClients;
-  failed_to_send?: QualifiedUserClients;
-  missing: QualifiedUserClients;
-  redundant: QualifiedUserClients;
-  time: string;
-}
+export * from './conversationApi';
