+++ libraries/api-client/src/conversation/index.ts
@@ -17,26 +17,26 @@
  *
  */
 
-export * from './Conversation';
-export * from './ConversationAPI';
-export * from './ConversationCode';
-export * from './ConversationError';
-export * from './FederatedBackendsError';
-export * from './ConversationIds';
-export * from './ConversationMembers';
-export * from './ConversationRole';
-export * from './Invite';
-export * from './Member';
-export * from './MessageSendingStatus';
-export * from './MutedStatus';
-export * from './NewConversation';
-export * from './NewOTRMessage';
-export * from './OtherMember';
-export * from './OTRClientMap';
-export * from './OTRRecipients';
-export * from './QualifiedUserClients';
-export * from './QualifiedConversationIds';
-export * from './RemoteConversations';
-export * from './ServiceRef';
-export * from './Subconversation';
-export * from './UserClients';
+export * from './conversation';
+export * from './conversationApi';
+export * from './conversationCode';
+export * from './conversationError';
+export * from './federatedBackendsError';
+export * from './conversationIds';
+export * from './conversationMembers';
+export * from './conversationRole';
+export * from './invite';
+export * from './member';
+export * from './messageSendingStatus';
+export * from './mutedStatus';
+export * from './newConversation';
+export * from './newOtrMessage';
+export * from './otherMember';
+export * from './otrClientMap';
+export * from './otrRecipients';
+export * from './qualifiedUserClients';
+export * from './qualifiedConversationIds';
+export * from './remoteConversations';
+export * from './serviceRef';
+export * from './subconversation';
+export * from './userClients';
