+++ libraries/api-client/src/conversation/qualifiedUserClients.ts
@@ -1,6 +1,6 @@
 /*
  * Wire
- * Copyright (C) 2022 Wire Swiss GmbH
+ * Copyright (C) 2021 Wire Swiss GmbH
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as published by
@@ -17,4 +17,8 @@
  *
  */
 
-export * from './ConversationAPI';
+import {UserClients} from './userClients';
+
+export interface QualifiedUserClients {
+  [domain: string]: UserClients;
+}
